5_Scripts/Pympo/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0        
ansys-api-mapdl==0.5.1  
ansys-mapdl-core==0.61.2
ansys-mapdl-reader==0.51.11
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
fonttools==4.31.2
frozenlist==1.3.0
geomdl==5.3.1
grpcio==1.44.0
idna==3.3
imageio==2.16.1
importlib-metadata==4.11.3
kiwisolver==1.4.2
matplotlib==3.5.1
multidict==6.0.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyiges==0.2.1
pyparsing==3.0.7
python-dateutil==2.8.2
pyvista==0.34.0
scipy==1.8.0
scooby==0.5.12
six==1.16.0
tqdm==4.63.1
vtk==9.1.0
wslink==1.5.0
yarl==1.7.2
zipp==3.7.0